requirements.txt
appium-python-client
pytest
selenium
